1)Completed implementation of NewBookingMarriage 2)Resolved error in getBookingStatus method in venue class

+++ TheFunctionJunction/GUI/src/main/java/Database/TableView/Drinks.java
@@ -25,7 +25,7 @@ public class Drinks extends Menu {
         this.myChoice = myChoice;
     }
     @Override
-    public float calculateCost() {
+    public double calculateCost() {
         return (float)(this.NumberOfGuests*this.PricePerPlate[this.getMyChoice()]);
     }
 }
